Renamings

+++ shared/src/blinkscript/script_built_ins.cpp
@@ -34,8 +34,8 @@ namespace SyncBlink
                                                    uint32_t countPerGroup = (uint32_t)frame.get("arg0").number;
                                                    for (uint32_t i = 0; i < blinkScript.getLed().getLedCount() / countPerGroup; i++)
                                                    {
-                                                       std::vector<uint32_t> ledIndices;
-                                                       for (uint32_t j = 0; j < countPerGroup; j++)
+                                                       std::vector<uint16_t> ledIndices;
+                                                       for (uint16_t j = 0; j < countPerGroup; j++)
                                                        {
                                                            ledIndices.push_back(i + (j * countPerGroup));
                                                        }
@@ -54,8 +54,8 @@ namespace SyncBlink
                         uint32_t groupCount = blinkScript.getLed().getLedCount()/countPerGroup;
                         for(uint32_t i = 0; i < groupCount; i++)
                         {
-                            std::vector<uint32_t> ledIndices;
-                            for(uint32_t j = 0; j < countPerGroup; j++)
+                            std::vector<uint16_t> ledIndices;
+                            for(uint16_t j = 0; j < countPerGroup; j++)
                             {
                                 if(j%2 == 0) ledIndices.push_back((i + j*groupCount));
 				                else ledIndices.push_back((j+1)*groupCount - (i+1));
@@ -108,7 +108,7 @@ namespace SyncBlink
                         if (arrayVal.getType() == ValueType::OBJECT && arrayVal.object->getType() == ObjectType::ARRAY)
                         {
                             auto arrayObj = static_cast<ArrayObj*>(arrayVal.object);
-                            for (int i = 0; i < arrayObj->getValues().size(); i++)
+                            for (uint32_t i = 0; i < arrayObj->getValues().size(); i++)
                             {
                                 blinkScript.getLed().setLed(i, Color::fromHex(arrayObj->getValues()[i].number));
                             }
